+++ server/src/main/scala/app.scala
@@ -20,7 +20,7 @@ import zio.logging.backend.SLF4J
 object MorbidServer extends GuaraApp {
 
   //FIXME: DO NOT ENABLE THIS. It breaks stack trace logging
-  //override val bootstrap = Runtime.removeDefaultLoggers >>> SLF4J.slf4j(LogFormat.colored)
+  override val bootstrap = Runtime.removeDefaultLoggers >>> SLF4J.slf4j //(LogFormat.colored)
 
   override val run = startGuara.provide(
     MorbidConfig.layer,