+++ changes.md
@@ -1,6 +1,7 @@
 # Morbid
 
  - Account/User provisioning (temporary)
+ - Enabled slf4j logging
 
 ## Release v0.0.3
 LTS: 01/11/2024