+++ changes.md
@@ -2,6 +2,7 @@
 
  - Account/User provisioning (temporary)
  - Enabled slf4j logging
+ - Added route POST `/app/:app/password/change`
 
 ## Release v0.0.3
 LTS: 01/11/2024